# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FitCore AI Fitness App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # ─── Users ──────────────────────────────────────────────────────────────────
...
  /exercises:
    get:
      operationId: listExercises
      tags: [exercises]
      summary: List exercises
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: muscleGroup
          in: query
          required: false
          schema:
            type: string
        - name: difficulty
          in: query
          required: false
          schema:
            type: string
            enum: [beginner, intermediate, advanced]
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
  /exercises/{id}:
    get:
      operationId: getExercise
      tags: [exercises]
      summary: Get an exercise
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "404":
          description: Not found

  # ─── Nutrition ───────────────────────────────────────────────────────────────
  /nutrition:
    get:
      operationId: listNutritionEntries
      tags: [nutrition]
      summary: List nutrition entries
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NutritionEntry"
    post:
      operationId: createNutritionEntry
      tags: [nutrition]
      summary: Log a meal or food item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NutritionEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionEntry"
  /nutrition/{id}:
    delete:
      operationId: deleteNutritionEntry
      tags: [nutrition]
      summary: Delete a nutrition entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /nutrition/summary:
    get:
      operationId: getNutritionSummary
      tags: [nutrition]
      summary: Get daily nutrition summary
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionSummary"

  # ─── Progress ────────────────────────────────────────────────────────────────
...
  /ai/conversations:
    get:
      operationId: listConversations
      tags: [ai]
      summary: List AI coach conversations
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [ai]
      summary: Start a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /ai/conversations/{id}/messages:
    get:
      operationId: getConversationMessages
      tags: [ai]
      summary: Get messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [ai]
      summary: Send a message to the AI coach
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ─── Subscriptions ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ─── Auth ──────────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
        name:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, name, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        subscriptionStatus:
          type: string
          enum: [free, premium, elite]
        onboardingCompleted:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ─── User Profile ──────────────────────────────────────────────────────────
...
    Exercise:
      type: object
      required: [id, name, category, difficulty, muscleGroups]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        muscleGroups:
          type: array
          items:
            type: string
        equipment:
          type: array
          items:
            type: string
        instructions:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        caloriesPerMinute:
          type: ["number", "null"]

    # ─── Nutrition ─────────────────────────────────────────────────────────────
    NutritionEntry:
      type: object
      required: [id, userId, name, calories, mealType, loggedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        calories:
          type: number
        protein:
          type: ["number", "null"]
        carbs:
          type: ["number", "null"]
        fat:
          type: ["number", "null"]
        fiber:
          type: ["number", "null"]
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        servingSize:
          type: ["string", "null"]
        loggedAt:
          type: string
          format: date-time
    NutritionEntryInput:
      type: object
      required: [name, calories, mealType]
      properties:
        name:
          type: string
        calories:
          type: number
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
        fiber:
          type: number
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        servingSize:
          type: string
        loggedAt:
          type: string
          format: date-time
    NutritionSummary:
      type: object
      required: [totalCalories, totalProtein, totalCarbs, totalFat, calorieGoal, entries]
      properties:
        totalCalories:
          type: number
        totalProtein:
          type: number
        totalCarbs:
          type: number
        totalFat:
          type: number
        calorieGoal:
          type: number
        entries:
          type: integer

    # ─── Progress ──────────────────────────────────────────────────────────────
...
    Conversation:
      type: object
      required: [id, userId, title, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        lastMessageAt:
          type: ["string", "null"]
          format: date-time
        messageCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    ConversationInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
    Message:
      type: object
      required: [id, conversationId, role, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # ─── Subscriptions ─────────────────────────────────────────────────────────
...