# lib/api-zod/src/generated/types/nutritionEntry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FitCore AI Fitness App API
 * OpenAPI spec version: 0.1.0
 */
import type { NutritionEntryMealType } from './nutritionEntryMealType';

export interface NutritionEntry {
  id: number;
  userId: number;
  name: string;
  calories: number;
  /** @nullable */
  protein?: number | null;
  /** @nullable */
  carbs?: number | null;
  /** @nullable */
  fat?: number | null;
  /** @nullable */
  fiber?: number | null;
  mealType: NutritionEntryMealType;
  /** @nullable */
  servingSize?: string | null;
  loggedAt: Date;
}