# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FitCore AI Fitness App API
 * OpenAPI spec version: 0.1.0
 */
import type { UserSubscriptionStatus } from './userSubscriptionStatus';

export interface User {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  subscriptionStatus?: UserSubscriptionStatus;
  onboardingCompleted?: boolean;
  createdAt: Date;
}